% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbart_new.R
\name{lbart_new}
\alias{lbart_new}
\title{Fit Bayesian Causal Forests}
\usage{
lbart_new(y, Omega, x_control, nburn = 1000, nsim = 1000, nthin = 1,
  update_interval = 100, ntree_control = 200, sd_control = 2 * sd(y),
  base_control = 0.95, power_control = 2, nu = 3, lambda = NULL,
  sigq = 0.9, sighat = NULL, include_pi = "control",
  use_muscale = FALSE)
}
\arguments{
\item{y}{Response variable}

\item{Omega}{Desgin matrix computed from treatment variable}

\item{x_control}{Design matrix for the function mu(x)}

\item{nburn}{Number of burn-in MCMC iterations}

\item{nsim}{Number of MCMC iterations to save after burn-in}

\item{nthin}{Save every nthin'th MCMC iterate. The total number of MCMC iterations will be nsim*nthin + nburn.}

\item{update_interval}{Print status every update_interval MCMC iterations}

\item{ntree_control}{Number of trees in mu(x)}

\item{sd_control}{SD(mu(x)) marginally at any covariate value (or its prior median if use_muscale=TRUE)}

\item{base_control}{Base for tree prior on mu(x) trees (see details)}

\item{power_control}{Power for the tree prior on mu(x) trees}

\item{nu}{Degrees of freedom in the chisq prior on \eqn{sigma^2}}

\item{lambda}{Scale parameter in the chisq prior on \eqn{sigma^2}}

\item{sigq}{Calibration quantile for the chisq prior on \eqn{sigma^2}}

\item{sighat}{Calibration estimate for the chisq prior on \eqn{sigma^2}}

\item{include_pi}{Takes values "control" or "none".}

\item{use_muscale}{Use a half-Cauchy hyperprior on the scale of mu.}

\item{pihat}{Length n estimates of}
}
\value{
A list with elements
\item{mu}{\code{nsim} by \code{n} matrix of posterior samples of individual treatment effects}
\item{sigma}{Length \code{nsim} vector of posterior samples of sigma}
}
\description{
Fit Bayesian Causal Forests
}
\details{
Fits a generalized version of the Bayesian Causal Forest model (Hahn et. al. 2018): For a response
variable y, treatment z, and covariates x,
\deqn{y_i = \mu(x_i, \hat z_i) + \tau(x_i, \pi_i)\omega(z_i) + \epsilon_i}
where \eqn{\z_i} is an (optional) estimate of \eqn{E(Z_i | X_i=x_i)} and
\eqn{\epsilon_i \sim N(0,\sigma^2)}

Some notes:
\itemize{
   \item x_control and x_moderate must be numeric matrices. See e.g. the makeModelMatrix function in the
   dbarts package for appropriately constructing a design matrix from a data.frame
   \item sd_control and sd_moderate are the prior SD(mu(x)) and SD(tau(x)) at a given value of x (respectively). If
   use_muscale = FALSE, then this is the parameter \eqn{\sigma_\mu} from the original BART paper, where the leaf parameters
   have prior distribution \eqn{N(0, \sigma_\mu/m)}, where m is the number of trees.
   If use_muscale=TRUE then sd_control is the prior median of a half Cauchy prior for SD(mu(x)). If use_tauscale = TRUE,
   then sd_moderate is the prior median of a half Normal prior for SD(tau(x)).
   \item By default the prior on \eqn{\sigma^2} is calibrated as in Chipman, George and McCulloch (2008).


}
}
\references{
Hahn, Murray, and Carvalho(2017). Bayesian regression tree models for causal inference: regularization, confounding, and heterogeneous effects.
 https://arxiv.org/abs/1706.09523. (Call citation("bcf") from the
command line for citation information in Bibtex format.)
}
